requirements.txt



fastapi==0.128.0
uvicorn==0.40.0
jinja2==3.1.6


tweepy==4.16.0
requests==2.32.5


python-dotenv==1.0.0


python-dateutil==2.8.2